overflow check

+++ demehin.maxim/S1/main.cpp
@@ -9,7 +9,7 @@
 namespace
 {
   //using ListOfPairs = std::list< std::pair< std::string, std::list< int > > >;
-  using ListOfPairs = demehin::List< std::pair< std::string, demehin::List< int > > >;
+  using ListOfPairs = demehin::List< std::pair< std::string, demehin::List< unsigned long long > > >;
 
   size_t defineMaxSize(ListOfPairs pairsList)
   {
@@ -21,9 +21,9 @@ namespace
     return max_size;
   }
 
-  int* copyNums(const int* nums, size_t nums_cnt)
+  unsigned long long* copyNums(const unsigned long long* nums, size_t nums_cnt)
   {
-    int* new_nums = new int[nums_cnt];
+    unsigned long long* new_nums = new unsigned long long[nums_cnt];
     for (size_t i = 0; i < nums_cnt; i++)
     {
       new_nums[i] = nums[i];
@@ -31,7 +31,7 @@ namespace
     return new_nums;
   }
 
-  void printSumValues(std::ostream& out, const int* vals, size_t size)
+  void printSumValues(std::ostream& out, const unsigned long long* vals, size_t size)
   {
     if (size > 0)
     {
@@ -50,13 +50,13 @@ namespace
   void printValues(std::ostream& out, ListOfPairs pairsList)
   {
     size_t capacity = 128;
-    int* sum_values = new int[capacity]();
+    unsigned long long* sum_values = new unsigned long long[capacity]();
     size_t max_size = defineMaxSize(pairsList);
     size_t sum_vals_cnt = 0;
-    int max_int = std::numeric_limits< int >::max();
+    unsigned long long max_int = std::numeric_limits< unsigned long long >::max();
     for (size_t i = 0; i < max_size; i++)
     {
-      int sum = 0;
+      unsigned long long sum = 0;
       bool isFirst = true;
       for (auto it = pairsList.begin(); it != pairsList.end(); it++)
       {
@@ -82,7 +82,7 @@ namespace
       if (i >= capacity)
       {
         capacity *= 2;
-        int* new_values = nullptr;
+        unsigned long long* new_values = nullptr;
         try
         {
           new_values = copyNums(sum_values, capacity);
@@ -107,18 +107,53 @@ namespace
 int main()
 {
   //std::list< std::pair< std::string, std::list< int > > > pairsList;
-  demehin::List< std::pair< std::string, demehin::List< int > > > pairsList;
+  demehin::List< std::pair< std::string, demehin::List< unsigned long long > > > pairsList;
   std::string nodeName;
   while (std::cin >> nodeName)
   {
-    demehin::List< int > numsList;
-    int num = 0;
+    demehin::List< unsigned long long > numsList;
+    unsigned long long num = 0;
+    //std::string num_str;
+    //int num = 0;
+    //std::getline(std::cin, num_str);
+    //while (std::cin >> num_str && std::cin.peek() != '\n')
     while (std::cin >> num)
     {
+      //if (num_str.empty())
+      //{
+        //b/reak;
+      //}
+      //try
+      //{
+        //unsigned long long num = std::stoll(num_str);
+        //if (len < num_str.length())
+        //{
+          //std::cerr << "erorr: overflow\n";
+          //return 1;
+        //}
+        //numsList.push_back(num);
+      //}
+      //catch (const std::out_of_range&)
+      //{
+        //std::cerr << "error: overflow\n";
+        //return 1;
+      //}
+      //catch (const std::invalid_argument&)
+     // {
+        //pairsList.push_back(std::make_pair(nodeName, numsList));
+      //}
+      //catch (const std::out_of_range&)
+      //{
+        //std::cerr << "error: overflow\n";
+        //return 1;
+      //}
+      //numsList.push_back(num);
       numsList.push_back(num);
     }
     pairsList.push_back(std::make_pair(nodeName, numsList));
     std::cin.clear();
+    //std::cin.ignore(std::numeric_limits< std::streamsize >::max(), '\n');
+    //std::cin.clear();
   }
 
   if (pairsList.size() == 0)
@@ -142,5 +177,6 @@ int main()
     std::cerr << e.what() << "\n";
     return 1;
   }
+
   std::cout << "\n";
 }
